% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_recent_data.R
\name{get_most_recent_data}
\alias{get_most_recent_data}
\title{Get Most Recent ILI data}
\usage{
get_most_recent_data(qry_data = FALSE)
}
\arguments{
\item{qry_data}{a logical flag (TRUE/FALSE) for whether to get data from API}
}
\value{
most recent ILI data
}
\description{
Get Most Recent ILI data
}
