% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_recent_data.R
\name{prc_most_recent_data}
\alias{prc_most_recent_data}
\title{Process Most Recent ILI Data}
\usage{
prc_most_recent_data(data)
}
\arguments{
\item{data}{raw ILI data}
}
\value{
most recent ILI data
}
\description{
This function gets the most recent ILI data
}
