% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_recent_data.R
\name{qry_data_from_api}
\alias{qry_data_from_api}
\title{Query Data from API}
\usage{
qry_data_from_api(limit = 2e+07)
}
\arguments{
\item{limit}{query limitations from API}
}
\value{
raw data for Influenza-like Illnesses (ILI)
}
\description{
This function gets Influenza-like Illnesses data
from NYC Open Data API.
}
